refactor(server): Move ApiValidators to zally-core #560
