+++ server/src/main/java/de/zalando/zally/configuration/RulesValidatorConfiguration.kt
@@ -1,10 +1,10 @@
 package de.zalando.zally.configuration
 
 import de.zalando.zally.core.DefaultContextFactory
-import de.zalando.zally.rule.ApiValidator
-import de.zalando.zally.rule.CompositeRulesValidator
-import de.zalando.zally.rule.ContextRulesValidator
-import de.zalando.zally.rule.JsonRulesValidator
+import de.zalando.zally.core.ApiValidator
+import de.zalando.zally.core.CompositeRulesValidator
+import de.zalando.zally.core.ContextRulesValidator
+import de.zalando.zally.core.JsonRulesValidator
 import de.zalando.zally.core.RulesManager
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.context.annotation.Bean
