+++ server/zally-core/src/main/kotlin/de/zalando/zally/core/JsonRulesValidator.kt
@@ -1,10 +1,7 @@
-package de.zalando.zally.rule
+package de.zalando.zally.core
 
 import com.fasterxml.jackson.core.JsonPointer
 import com.fasterxml.jackson.databind.JsonNode
-import de.zalando.zally.core.ContentParseResult
-import de.zalando.zally.core.ObjectTreeReader
-import de.zalando.zally.core.RulesManager
 import de.zalando.zally.core.ast.ReverseAst
 import io.swagger.util.Json
